# Introduce %EXPANSION, optimize LET-SYMBOL